transients: Search for both client and window frame
Bug: 12206

Android Virtual Device emulator marks its utility window as a transient,
but xfwm4 fails to to match it against the emulator window, which breaks
the transient relationship.

As the AVD emulator raises its utility window, it grabs focus and causes
sort of problems.

Look for both window and frame windows when matching transients.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>